cleaned up warnings and started with the logging the llx/I64x format specifier seems to throw a warning no matter what on mingw :(

+++ main/Architecture.cpp
@@ -36,7 +36,7 @@ namespace holodec {
 			HId id = entry.first;
 			instrIds.insert(id);
 			IRParser parser (this);
-			for (int i = 0; i < entry.second.irs.size(); i++) {
+			for (size_t i = 0; i < entry.second.irs.size(); i++) {
 				if (entry.second.irs[i]) {
 					parser.parse (&entry.second.irs[i]);
 				}
